Updated Webinterface and XML creation and parsing.
Replaced the "Rebuild XML" and "Reload Table" buttons with the "Rebuild
XML and Reload Table" button which executes a script that waits for the
php-script to finish and reloads the table afterwards.
Modified and fixed Timestamp creation (createXMl) and parsing
(Webinterface).
Changed XML creation and parsing:
Removed Nodes for single entry information (ID, ConnectionCount and the
Connected flag).