Added logignore compatibility.
Added possibility to place a "logignore" in the logdirectory.  The logs
specified (by filename) in this file won't be parsed.
Very helpful after a backup was used or when a database migration went
wrong.
Added button for a complete rebuild of the XML just from parsing the
logs to the Webinterface (useful if switching logdirectories).