Huge update - Added kick table.
Added the classes Kick and Ban for storing kicks and bans - at the
moment only Kick is useable.
Adjusted parsing of logs and XML as well as the creation of the XML
according to the new kick class.
Updated the makefile.
Fixed a problem with the "Back to top"-button.


Former-commit-id: 3ffa54bbdd02ba4bde526faf3c3e737e33f1121f
Former-commit-id: aa3d12390479eec02614d21aaddab91aaeecfbd8