Huge update: added table for uploaded files.
Added the class File and updated the regular features including parsing,
XML creation and so on (take a look at the diffs for all the changes).
Added notes for the implemention of the ban coverage but postponed this
part of the development.
Updated compiled files to the current code version.


Former-commit-id: c5820d4e962d9ac38f4e7cc9a9c1c3f4fc50d5b1
Former-commit-id: b5af9a58941e8b3d8917dbf382a530f1ee4cce73