Rewrote js code and made interface look more modern
Rewrote the whole tablebuilding javascript code in order to provide a
clearer and less error-prone code.
Moved the style.css include and most of the javascript includes from the
"index.html" to the "TS3_EnhancedClientList.js" in order to provide an
easier use with custom pages.
Prevented multiple php-requests at a time by disabling the rebuild
buttons until the current request and the following table building are
finished.
The XML isn't automatically rebuild anymore when the page is reloaded
but only when one of the rebuild buttons is used.
Changed original tablesorter version to this fork
(http://mottie.github.io/tablesorter/docs/) as it provides new features
like sticky headers.
Many changes to the "style.css" in order to provide a more modern style
(still much more to come) and to match the new tablesorter version.
Added sticky headers to the tables.
Modified the "scrollToDiv" function and added the new function
"scrollToClientTableRow" which handles the offset of the sticky headers
when scrolling to an ID in the client table.
Added moment.js and included it into the control section, the kick table
and the upload table.
Added the custom table parser "ignoreMoment" to make sorting of the
tables containing the moment.js timestamps possible.
Replaced the two buttons for expanding and collapsing the lists with one
button that expands or collapses the current list according to its
current state.
Added a loading spinner which is displayed while the XML and the tables
are rebuild.
Updated README and Install Instructions (as usual).


Former-commit-id: e2b0b68eaff388b5097f1f0deb565673b292c082
Former-commit-id: eac621ef2152ad88ab03e40966754815c0f1e681