Minor code improvements + modified XML creation.
Removed unnecessary checks from methods of the user class and moved one
necessary check into the parseLogs.cpp.
Moved the attributes to the beginning of the XML and included the
virtualserver in the attributes section. The XML is now ignored if the
current and the virtual server in the XML differ.
Moved the definitions of the default logdirectory and the default
virtual server into the Constants.h.


Former-commit-id: 38e4eed1c82af765a3693a3177034e740a14ba21
Former-commit-id: a0d2143425c38f7614a8f3de6edd1bcc42386c4d