Extended Ban parsing and several renamings.
Added parsing for the bannedUID, bannedIP and bannedByIP and added them
to the Ban class while removing the bannedByInvokerID.
Changed the beginning of the check functions name from uppercase to
lowercase (Is -> is).
Renamed complaintForNickname to complaintAboutNickname and
complaintForID to complaintAboutID.
Updated Table headlines and updated the readme according to this
changes.
Surrounded the xml parsing with a try catch block in order to prevent
problems when changing to a newer version with changed xml parameters.


Former-commit-id: 6231768741c9aa03e8c9af4349c28fd0f3f28bd0
Former-commit-id: abfc29d805e22dc192889dfb923298c6b3d88a9e