Style rework using foundation.
Style rework using the front-end framework foundation.
Moved controls for certain tables to the top of this certain tables.
Replaced the switch for switching between the sorting by the first or
the last connect with a button.
Removed the "Scroll to ID in the client list"-button.
Fixed the problem with expanding connections when the total count of
connections was three.
Updated install instructions and readme.


Former-commit-id: 959181376d3a83b19646c96c12052d58a2581088
Former-commit-id: 6d73436f59bf699f4c05700c1ef06e009e159d86