Fixed two issues and modified Upload deleted flags.
The deleted flags for Uploads are now only written to the XML if the
upload was deleted.
Fixed parsing for kicks where the client was kicked by a serveradmin (=
no real UID).
Fixed switching between IDs and UIDs for the ban table.
Outsourced the functions adding the onclick functions out of
buildClientTable().


Former-commit-id: 50f9e4a3cd0f2317dd6b9dae97b55bd05badeac6
Former-commit-id: ceb1d391033991176955daaf326d2becd2eb5a8a